[#51] enhancement: accept double value in zoom by method (#58)
* [#51] make changes in zoom by method to accept double value

* [#51] formatted file

* [#51] updated change logs

+++ android/src/main/kotlin/com/itheamc/naxalibre/NaxaLibreController.kt
@@ -466,8 +466,8 @@ class NaxaLibreController(
      * Zooms the camera by a specified amount.
      * @param by The amount to zoom by.
      */
-    override fun zoomBy(by: Long) {
-        val cameraUpdate = CameraUpdateFactory.zoomBy(by.toDouble())
+    override fun zoomBy(by: Double) {
+        val cameraUpdate = CameraUpdateFactory.zoomBy(by)
         libreMap.animateCamera(cameraUpdate)
     }
